pyproject.toml
[tool.poetry]
name = "toyshop"
version = "0.1.0"
description = "RESTful API for online toy shop"
authors = ["Sijan Shrestha <sijan.shrestha@cloco.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
flask-marshmallow = "^1.2.1"
sqlalchemy = "^2.0.36"
flask-sqlalchemy = "^3.1.1"
marshmallow-sqlalchemy = "^1.1.0"
flask-migrate = "^4.0.7"
flask-jwt-extended = "^4.7.1"
flask-bcrypt = "^1.0.1"
flask-smorest = "^0.45.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
flask-swagger-ui = "^4.11.1"
email-validator = "^2.2.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
ruff = "^0.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
